Convert Big Decimal and Float to Double for Work

+++ app/src/main/java/pro/selecto/slothos/data/model/MeasurementBase.kt
@@ -1,13 +1,11 @@
 package pro.selecto.slothos.data.model
 
-import java.math.BigDecimal
-
 interface MeasurementBase {
     val id: Int
     val name: String
     val description: String
     val symbol: String
-    val conversionToBase: BigDecimal
+    val conversionToBase: Double
     val categoryType: MeasurementCategory
     val defaultPrecision: Int
 }
